MyProjFolder/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues


azure-functions

requests==2.32.3
bs4==0.0.2